HW/HW2/fashion_mlops/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
Pillow>=10.2.0
torch>=2.2.0
torchvision>=0.17.0
transformers==4.40.2
einops>=0.7.0
